+++ plone/formwidget/geolocation/__init__.py
@@ -2,4 +2,4 @@
 from plone.formwidget.geolocation.field import GeolocationField
 from plone.formwidget.geolocation.geolocation import Geolocation
 
-__all__ = ('Geolocation', 'GeolocationField')
+__all__ = ("Geolocation", "GeolocationField")