+++ plone/formwidget/geolocation/setuphandlers.py
@@ -5,9 +5,8 @@ from zope.interface import implementer
 
 @implementer(INonInstallable)
 class HiddenProfiles(object):
-
     def getNonInstallableProfiles(self):
         """Hide uninstall profile from site-creation and quickinstaller"""
         return [
-            'plone.formwidget.geolocation:uninstall',
+            "plone.formwidget.geolocation:uninstall",
         ]