+++ plone/formwidget/geolocation/vocabularies.py
@@ -8,12 +8,12 @@ from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
 
 
-_ = MessageFactory('plone.formwidget.geolocation')
-default_map_layer = 'OpenStreetMap.Mapnik'
+_ = MessageFactory("plone.formwidget.geolocation")
+default_map_layer = "OpenStreetMap.Mapnik"
 default_map_layers = [
-    'OpenStreetMap.Mapnik',
-    'Esri.WorldImagery',
-    'CartoDB.DarkMatter',
+    "OpenStreetMap.Mapnik",
+    "Esri.WorldImagery",
+    "CartoDB.DarkMatter",
 ]
 
 
@@ -25,36 +25,108 @@ def MapLayers(context):
     """
     request = getRequest()
     items = [
-        (_('OpenStreetMap.Mapnik',           default='OpenStreetMap Mapnik'),           'OpenStreetMap.Mapnik'),            # noqa
-        (_('OpenStreetMap.BlackAndWhite',    default='OpenStreetMap Black & White'),    'OpenStreetMap.BlackAndWhite'),     # noqa
-        (_('OpenStreetMap.DE',               default='OpenStreetMap DE'),               'OpenStreetMap.DE'),                # noqa
-        (_('OpenStreetMap.France',           default='OpenStreetMap France'),           'OpenStreetMap.France'),            # noqa
-        (_('Thunderforest.OpenCycleMap',     default='Thunderforest Cycle Map'),        'Thunderforest.OpenCycleMap'),      # noqa
-        (_('Thunderforest.Transport',        default='Thunderforest Transport'),        'Thunderforest.Transport'),         # noqa
-        (_('Thunderforest.TransportDark',    default='Thunderforest Transport Dark'),   'Thunderforest.TransportDark'),     # noqa
-        (_('Thunderforest.Outdoors',         default='Thunderforest Outdoors'),         'Thunderforest.Outdoors'),          # noqa
-        (_('Stamen.Toner',                   default='Stamen Toner'),                   'Stamen.Toner'),                    # noqa
-        (_('Stamen.TonerBackground',         default='Stamen Toner Background'),        'Stamen.TonerBackground'),          # noqa
-        (_('Stamen.TonerLite',               default='Stamen Toner Lite'),              'Stamen.TonerLite'),                # noqa
-        (_('Stamen.Watercolor',              default='Stamen Watercolor'),              'Stamen.Watercolor'),               # noqa
-        (_('Stamen.Terrain',                 default='Stamen Terrain'),                 'Stamen.Terrain'),                  # noqa
-        (_('Stamen.TerrainBackground',       default='Stamen Terrain Background'),      'Stamen.TerrainBackground'),        # noqa
-        (_('Stamen.TopOSMRelief',            default='Stamen Relief'),                  'Stamen.TopOSMRelief'),             # noqa
-        (_('Esri.WorldStreetMap',            default='Esri World Street Map'),          'Esri.WorldStreetMap'),             # noqa
-        (_('Esri.DeLorme',                   default='Esri DeLorme'),                   'Esri.DeLorme'),                    # noqa
-        (_('Esri.WorldTopoMap',              default='Esri World Topo Map'),            'Esri.WorldTopoMap'),               # noqa
-        (_('Esri.WorldImagery',              default='Esri World Imagery'),             'Esri.WorldImagery'),               # noqa
-        (_('Esri.WorldTerrain',              default='Esri World Terrain'),             'Esri.WorldTerrain'),               # noqa
-        (_('Esri.WorldShadedRelief',         default='Esri World Shaded Relief'),       'Esri.WorldShadedRelief'),          # noqa
-        (_('Esri.WorldPhysical',             default='Esri World Physical'),            'Esri.WorldPhysical'),              # noqa
-        (_('Esri.OceanBasemap',              default='Esri Ocean Basemap'),             'Esri.OceanBasemap'),               # noqa
-        (_('Esri.NatGeoWorldMap',            default='Esri Nat Geo World Map'),         'Esri.NatGeoWorldMap'),             # noqa
-        (_('Esri.WorldGrayCanvas',           default='Esri World Gray Canvas'),         'Esri.WorldGrayCanvas'),            # noqa
-        (_('CartoDB.DarkMatter',             default='CatoDB Dark Matter'),             'CartoDB.DarkMatter'),              # noqa
-        (_('CartoDB.DarkMatterNoLabels',     default='CartoDB Dark Matter No Labels'),  'CartoDB.DarkMatterNoLabels'),      # noqa
-        (_('NASAGIBS.ViirsEarthAtNight2012', default='NASA Earth at Night 2012'),       'NASAGIBS.ViirsEarthAtNight2012'),  # noqa
+        (
+            _("OpenStreetMap.Mapnik", default="OpenStreetMap Mapnik"),
+            "OpenStreetMap.Mapnik",
+        ),
+        (
+            _("OpenStreetMap.BlackAndWhite", default="OpenStreetMap Black & White"),
+            "OpenStreetMap.BlackAndWhite",
+        ),
+        (_("OpenStreetMap.DE", default="OpenStreetMap DE"), "OpenStreetMap.DE"),
+        (
+            _("OpenStreetMap.France", default="OpenStreetMap France"),
+            "OpenStreetMap.France",
+        ),
+        (
+            _("Thunderforest.OpenCycleMap", default="Thunderforest Cycle Map"),
+            "Thunderforest.OpenCycleMap",
+        ),
+        (
+            _("Thunderforest.Transport", default="Thunderforest Transport"),
+            "Thunderforest.Transport",
+        ),
+        (
+            _("Thunderforest.TransportDark", default="Thunderforest Transport Dark"),
+            "Thunderforest.TransportDark",
+        ),
+        (
+            _("Thunderforest.Outdoors", default="Thunderforest Outdoors"),
+            "Thunderforest.Outdoors",
+        ),
+        (_("Stamen.Toner", default="Stamen Toner"), "Stamen.Toner"),
+        (
+            _("Stamen.TonerBackground", default="Stamen Toner Background"),
+            "Stamen.TonerBackground",
+        ),
+        (
+            _("Stamen.TonerLite", default="Stamen Toner Lite"),
+            "Stamen.TonerLite",
+        ),
+        (
+            _("Stamen.Watercolor", default="Stamen Watercolor"),
+            "Stamen.Watercolor",
+        ),
+        (_("Stamen.Terrain", default="Stamen Terrain"), "Stamen.Terrain"),
+        (
+            _("Stamen.TerrainBackground", default="Stamen Terrain Background"),
+            "Stamen.TerrainBackground",
+        ),
+        (
+            _("Stamen.TopOSMRelief", default="Stamen Relief"),
+            "Stamen.TopOSMRelief",
+        ),
+        (
+            _("Esri.WorldStreetMap", default="Esri World Street Map"),
+            "Esri.WorldStreetMap",
+        ),
+        (_("Esri.DeLorme", default="Esri DeLorme"), "Esri.DeLorme"),
+        (
+            _("Esri.WorldTopoMap", default="Esri World Topo Map"),
+            "Esri.WorldTopoMap",
+        ),
+        (
+            _("Esri.WorldImagery", default="Esri World Imagery"),
+            "Esri.WorldImagery",
+        ),
+        (
+            _("Esri.WorldTerrain", default="Esri World Terrain"),
+            "Esri.WorldTerrain",
+        ),
+        (
+            _("Esri.WorldShadedRelief", default="Esri World Shaded Relief"),
+            "Esri.WorldShadedRelief",
+        ),
+        (
+            _("Esri.WorldPhysical", default="Esri World Physical"),
+            "Esri.WorldPhysical",
+        ),
+        (
+            _("Esri.OceanBasemap", default="Esri Ocean Basemap"),
+            "Esri.OceanBasemap",
+        ),
+        (
+            _("Esri.NatGeoWorldMap", default="Esri Nat Geo World Map"),
+            "Esri.NatGeoWorldMap",
+        ),
+        (
+            _("Esri.WorldGrayCanvas", default="Esri World Gray Canvas"),
+            "Esri.WorldGrayCanvas",
+        ),
+        (
+            _("CartoDB.DarkMatter", default="CatoDB Dark Matter"),
+            "CartoDB.DarkMatter",
+        ),
+        (
+            _("CartoDB.DarkMatterNoLabels", default="CartoDB Dark Matter No Labels"),
+            "CartoDB.DarkMatterNoLabels",
+        ),
+        (
+            _("NASAGIBS.ViirsEarthAtNight2012", default="NASA Earth at Night 2012"),
+            "NASAGIBS.ViirsEarthAtNight2012",
+        ),
     ]
     items = [
-        SimpleTerm(title=translate(i[0], context=request), value=i[1])
-        for i in items]
+        SimpleTerm(title=translate(i[0], context=request), value=i[1]) for i in items
+    ]
     return SimpleVocabulary(items)