+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - Allow to set default geolocation on new contents (via bool field in config).
   If not checked, the defaut geolocation is only used to center map.
+  Also, the geolocation map will not show on an object if no geolocation was defined.
   [laulaz]